# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FitForge Workout App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workouts:
    get:
      operationId: listWorkouts
      tags: [workouts]
      summary: List workout plans
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workout"
    post:
      operationId: createWorkout
      tags: [workouts]
      summary: Create workout plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
  /workouts/{id}:
    get:
      operationId: getWorkout
      tags: [workouts]
      summary: Get workout plan with exercises
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutDetail"
        "404":
          description: Not found
    patch:
      operationId: updateWorkout
      tags: [workouts]
      summary: Update workout plan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
    delete:
      operationId: deleteWorkout
      tags: [workouts]
      summary: Delete workout plan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Workout Logs ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Workout:
      type: object
      required: [id, name, difficulty, estimatedMinutes, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        difficulty:
          type: string
        estimatedMinutes:
          type: integer
        category:
          type: ["string", "null"]
        isAiGenerated:
          type: boolean
        exerciseCount:
          type: integer
        createdAt:
          type: string
    WorkoutInput:
      type: object
      required: [name, difficulty, estimatedMinutes]
      properties:
        name:
          type: string
        description:
          type: string
        difficulty:
          type: string
        estimatedMinutes:
          type: integer
        category:
          type: string
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutExerciseInput"
    WorkoutUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        difficulty:
          type: string
        estimatedMinutes:
          type: integer
        category:
          type: string
    WorkoutExerciseInput:
      type: object
      required: [exerciseId, sets, reps]
      properties:
        exerciseId:
          type: integer
        sets:
          type: integer
        reps:
          type: integer
        weight:
          type: ["number", "null"]
        restSeconds:
          type: integer
        notes:
          type: string
        orderIndex:
          type: integer
    WorkoutExercise:
      type: object
      required: [id, exerciseId, sets, reps, orderIndex]
      properties:
        id:
          type: integer
        exerciseId:
          type: integer
        exerciseName:
          type: string
        muscleGroup:
          type: string
        equipment:
          type: string
        sets:
          type: integer
        reps:
          type: integer
        weight:
          type: ["number", "null"]
        restSeconds:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        orderIndex:
          type: integer
    WorkoutDetail:
      type: object
      required: [id, name, difficulty, estimatedMinutes, createdAt, exercises]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        difficulty:
          type: string
        estimatedMinutes:
          type: integer
        category:
          type: ["string", "null"]
        isAiGenerated:
          type: boolean
        createdAt:
          type: string
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutExercise"

    # ── Workout Logs ──────────────────────────────────────────────────────
...