# lib/api-zod/src/generated/types/workout.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitForge Workout App API
 * OpenAPI spec version: 0.1.0
 */

export interface Workout {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  difficulty: string;
  estimatedMinutes: number;
  /** @nullable */
  category?: string | null;
  isAiGenerated?: boolean;
  exerciseCount?: number;
  createdAt: string;
}