# lib/api-zod/src/generated/types/workoutDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitForge Workout App API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutExercise } from './workoutExercise';

export interface WorkoutDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  difficulty: string;
  estimatedMinutes: number;
  /** @nullable */
  category?: string | null;
  isAiGenerated?: boolean;
  createdAt: string;
  exercises: WorkoutExercise[];
}